+++ web/src/lib/images.ts
@@ -1,5 +1,6 @@
 import { buildUrl } from 'cloudinary-build-url'
-import { CLOUDINARY_CLOUD_NAME } from 'lib/constants'
+
+import { CLOUDINARY_CLOUD_NAME } from 'src/lib/constants'
 
 export const cloudinaryUrl = (
   id: string,
